Moved representations and request parameters from Event Module
